@@ -1,31 +1,31 @@
 name: stream_chat
 homepage: https://getstream.io/
 description: The official Dart client for Stream Chat, a service for building chat applications.
-version: 7.2.2
+version: 8.0.0
 repository: https://github.com/GetStream/stream-chat-flutter
 issue_tracker: https://github.com/GetStream/stream-chat-flutter/issues
 publish_to: 'none'
 
 environment:
-  sdk: '>=3.1.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   async: ^2.11.0
   collection: ^1.17.2
-  dio: ^5.3.2
+  dio: ^5.4.3+1
   equatable: ^2.0.5
   freezed_annotation: ^2.4.1
   http_parser: ^4.0.2
   jose: ^0.3.4
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
   logging: ^1.2.0
   meta: ^1.9.1
-  mime: ^1.0.4
+  mime: ^1.0.5
   rate_limiter: ^1.0.0
   rxdart: ^0.27.7
-  synchronized: ^3.1.0
-  uuid: ^4.2.1
-  web_socket_channel: ^2.4.0
+  synchronized: ^3.1.0+1
+  uuid: ^4.4.0
+  web_socket_channel: ^2.4.5
 
 dev_dependencies:
   build_runner: ^2.4.6
