@@ -4,8 +4,8 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.1.0 <4.0.0"
-  flutter: ">=3.13.0"
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ">=3.19.0"
 
 dependencies:
   collection: ^1.15.0
@@ -13,9 +13,9 @@ dependencies:
   flutter:
     sdk: flutter
   responsive_builder: ^0.7.0
-  stream_chat_flutter: ^7.2.2
-  stream_chat_localizations: ^7.2.2
-  stream_chat_persistence: ^7.2.2
+  stream_chat_flutter: ^8.0.0
+  stream_chat_localizations: ^8.0.0
+  stream_chat_persistence: ^8.0.0
 
 flutter:
   uses-material-design: true
