@@ -5,15 +5,15 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.1.0 <4.0.0"
-  flutter: ">=3.13.0"
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ">=3.19.0"
 
 dependencies:
   cupertino_icons: ^1.0.3
   flutter:
     sdk: flutter
-  stream_chat_flutter: ^7.2.2
-  stream_chat_localizations: ^7.2.2
+  stream_chat_flutter: ^8.0.0
+  stream_chat_localizations: ^8.0.0
 
 flutter:
   uses-material-design: true
