@@ -838,12 +838,11 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
           title: Text(context.translations.copyMessageLabel),
           onClick: () {
             Navigator.of(context, rootNavigator: true).pop();
-            final text = widget.message.text;
-            String? messageToCopy;
+            var messageToCopy = widget.message.text;
             for (final user in widget.message.mentionedUsers.toSet()) {
               final userId = user.id;
               final userName = user.name;
-              messageToCopy = text?.replaceAll(
+              messageToCopy = messageToCopy?.replaceAll(
                 RegExp('@($userId|$userName)'),
                 '@$userName',
               ) ??
@@ -1076,12 +1075,11 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
                   : DisplayWidget.show,
             ),
             onCopyTap: (message) {
-              final text = message.text;
-              String? messageToCopy;
+              var messageToCopy = message.text;
               for (final user in widget.message.mentionedUsers.toSet()) {
                 final userId = user.id;
                 final userName = user.name;
-                messageToCopy = text?.replaceAll(
+                messageToCopy = messageToCopy?.replaceAll(
                   RegExp('@($userId|$userName)'),
                   '@$userName',
                 ) ??
