@@ -75,7 +75,7 @@ void main() {
         StreamChat(
           streamChatThemeData: theme,
           client: client,
-          connectivityStream: Stream.value(ConnectivityResult.mobile),
+          connectivityStream: Stream.value([ConnectivityResult.mobile]),
           child: StreamChannel(
             showLoading: false,
             channel: channel,
@@ -136,7 +136,7 @@ void main() {
           home: StreamChat(
             streamChatThemeData: theme,
             client: client,
-            connectivityStream: Stream.value(ConnectivityResult.mobile),
+            connectivityStream: Stream.value([ConnectivityResult.mobile]),
             child: StreamChannel(
               showLoading: false,
               channel: channel,
@@ -194,7 +194,7 @@ void main() {
           home: StreamChat(
             streamChatThemeData: theme,
             client: client,
-            connectivityStream: Stream.value(ConnectivityResult.mobile),
+            connectivityStream: Stream.value([ConnectivityResult.mobile]),
             child: StreamChannel(
               showLoading: false,
               channel: channel,
