@@ -25,7 +25,7 @@ void main() {
         StreamChat(
           client: client,
           streamChatThemeData: theme,
-          connectivityStream: Stream.value(ConnectivityResult.mobile),
+          connectivityStream: Stream.value([ConnectivityResult.mobile]),
           child: Scaffold(
             body: Center(
               child: StreamReactionBubble(
@@ -68,7 +68,7 @@ void main() {
         StreamChat(
           client: client,
           streamChatThemeData: StreamChatThemeData.fromTheme(themeData),
-          connectivityStream: Stream.value(ConnectivityResult.mobile),
+          connectivityStream: Stream.value([ConnectivityResult.mobile]),
           child: ColoredBox(
             color: Colors.black,
             child: StreamReactionBubble(
@@ -105,7 +105,7 @@ void main() {
         StreamChat(
           client: client,
           streamChatThemeData: StreamChatThemeData.fromTheme(themeData),
-          connectivityStream: Stream.value(ConnectivityResult.mobile),
+          connectivityStream: Stream.value([ConnectivityResult.mobile]),
           child: ColoredBox(
             color: Colors.black,
             child: StreamReactionBubble(
@@ -150,7 +150,7 @@ void main() {
         StreamChat(
           client: client,
           streamChatThemeData: StreamChatThemeData.fromTheme(themeData),
-          connectivityStream: Stream.value(ConnectivityResult.mobile),
+          connectivityStream: Stream.value([ConnectivityResult.mobile]),
           child: ColoredBox(
             color: Colors.black,
             child: StreamReactionBubble(
@@ -195,7 +195,7 @@ void main() {
       await tester.pumpWidgetBuilder(
         StreamChat(
           client: client,
-          connectivityStream: Stream.value(ConnectivityResult.mobile),
+          connectivityStream: Stream.value([ConnectivityResult.mobile]),
           streamChatThemeData: StreamChatThemeData.fromTheme(themeData),
           child: Scaffold(
             body: Center(
